Removed client-admin dependency from function-utils
